Fixes documentation build errors
When using the nuget sandcastle packages, the documentation wouldn't
build remarks or code examples. Also, if you did not have the full
sandcastle installed, VS wouldnt see it and wouldn't build it anyway, so
there was not much of a real reason to have it in the solution. This
removed it from the solution, added a batch script in the root that will
build the documentation and grab the nuget packages if people want to
build the xml docs. However normal builds will not anymore.

Fixes Appveyor yml

Bad formatting

Changes AppVeyor to release mode

Appveyor should now build nupkg after documentation

Should now properly publish nupkg

Makes AppVeyor configuration build documentation

This fixes the issue where nugets would get packed before the post build
script runs. This also means that if the user runs the build in appveyor
mode it will download the nuget packages and build the docs.

Updates after target name