Always uses a unique temp file
The reason I originally implemented keeping the same file was that the
original ntcore library was over 7mb. Its now down to 500kb, so its
quicker to just extract a new copy. So reverted to just always using a
unique temp file for the library. With the hooks to delete the file on
unload, this should be fine.